Extracts page from name form
Also renames the presentation form for legal name to match the
container.
